requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
httpx>=0.27.0
google-generativeai>=0.3.2
mcp>=1.0.0
python-dotenv>=1.0.0
rich>=13.7.0 